WRK25USERS-8 Created PasswordFactory class with test, refactored PasswordStrengthValidator and tests to not throw exceptions
